+++ packages/hooks/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @youversion/platform-react-hooks
 
+## 2.7.1
+
+### Patch Changes
+
+- Updated dependencies [4f4ce4a]
+  - @youversion/platform-core@2.7.1
+
+## 2.7.0
+
+### Patch Changes
+
+- @youversion/platform-core@2.7.0
+
 ## 2.6.3
 
 ### Patch Changes